refactor(api): reorganizar estructura de API en pixela-frontend y modularizar configuración
Este commit reorganiza la estructura de la API para hacerla más modular y mantenible:

- Mover archivos de configuración de API de /config a /api/shared:
  - api.ts → api/shared/api.ts
  - apiEndpoints.ts → api/shared/apiEndpoints.ts
  - apiHelpers.ts → api/shared/apiHelpers.ts
  - apiTypes.ts → api/shared/apiTypes.ts

- Crear nueva estructura modular de API:
  - /api/auth/: Endpoints y tipos relacionados con autenticación
  - /api/favorites/: Gestión de favoritos
  - /api/reviews/: Funcionalidad de reseñas
  - /api/shared/: Utilidades y configuración compartida
  - /api/users/: Gestión de usuarios

- Eliminar archivos obsoletos:
  - /lib/api/index.ts
  - /lib/api/series/series.ts
  - /lib/index.ts
  - /lib/interface/movies/trending-movies.ts
  - /lib/interface/series/trending-series.ts

- Actualizar importaciones en archivos afectados:
  - /api/peliculas/mapPelicula.ts
  - /api/peliculas/peliculas.ts
  - /api/series/series.ts
  - /data/index.ts
  - /features/profile/components/layout/*
  - /features/profile/pages/ProfilePage.tsx
  - /stores/useAuthStore.ts

Esta refactorización mejora la organización del código al:
- Centralizar todo el código relacionado con la API en un único directorio
- Separar las responsabilidades en módulos lógicos
- Hacer el código más mantenible y escalable
- Reducir la duplicación y mejorar la reutilización del código
