+++ pixela-frontend/src/api/series/series.ts
@@ -1,7 +1,7 @@
 import { Actor } from '@/features/media/types/people';
 import { Serie } from '@/features/media/types/content';
-import { API_ENDPOINTS } from '@/config/apiEndpoints';
-import { DEFAULT_FETCH_OPTIONS } from '@/config/api';
+import { API_ENDPOINTS } from '../shared/apiEndpoints';
+import { DEFAULT_FETCH_OPTIONS } from '../shared/apiHelpers';
 import { mapSerieFromApi } from './mapSerie';
 
 export async function getSerieById(id: string): Promise<Serie> {
