+++ pixela-frontend/src/api/shared/apiEndpoints.ts
