+++ pixela-frontend/src/data/index.ts
@@ -1,2 +1,1 @@
-export * from "./hero/hero";
 export * from "./links/navigation";
