--- pixela-frontend/src/lib/index.ts
@@ -1,2 +1,0 @@
-export * from './api';
-export * from './interface/series/trending-series';
