--- pixela-frontend/src/lib/interface/series/trending-series.ts
@@ -1,12 +1,0 @@
-export interface SeriesResponse {
-    success: boolean;
-    data: Series[];
-}
-
-export interface Series {
-    id: number;
-    title: string;
-    poster_path: string;
-    first_air_date: string;
-    vote_average: number;
-}
